make price int

+++ src/api/bottler.py
@@ -35,7 +35,7 @@ def post_deliver_bottles(potions_delivered: list[PotionInventory], order_id: int
             green_ml += type[1] * qty
             blue_ml += type[2] * qty
             dark_ml += type[3] * qty
-            price = red_ppm*type[0] + green_ppm*type[1] + blue_ppm*type[2] + dark_ppm*type[3]
+            price = int(red_ppm*type[0] + green_ppm*type[1] + blue_ppm*type[2] + dark_ppm*type[3])
             sku = "R" + str(type[0]) + "_G" + str(type[1]) + "_B" + str(type[2]) + "_D" + str(type[3])
 
             result = connection.execute(sqlalchemy.text("SELECT sku FROM potions WHERE sku = :sku"),
